requirements.txt
Flask
Flask-Caching
requests
yt-dlp
google-generativeai
python-dotenv
psutil